requirements.txt
# Existing Flask dependencies
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6
requests==2.31.0
python-dateutil==2.8.2
jsonschema==4.19.1
pytest==7.4.2
pytest-flask==1.2.0

# New for RAG and Hugging Face
transformers==4.30.2
sentence-transformers==2.2.2
torch==2.0.1
faiss-cpu==1.7.4  # Or faiss-gpu if you have a compatible GPU

# PDF RAG System Dependencies (LangChain + Astra DB)
cassio>=0.1.0
langchain>=0.0.350
openai>=1.0.0
tiktoken>=0.5.0
PyPDF2>=3.0.0

# New for Production Features
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
pypdf==4.0.1
python-docx==1.1.0

# Additional utilities for document processing
markdown==3.5.1
beautifulsoup4==4.12.2
nltk==3.8.1
